wildfly: structures: fix beam ammount parameter type and improve simulation

+++ src/main/java/io/github/ipl/tfc/docetressabores/entities/structures/LightSteelStructure.java
@@ -29,7 +29,7 @@ public class LightSteelStructure extends Structure {
 		Material material,
 		String name,
 		int beamAmount,
-		int beamLength,
+		double beamLength,
 		int beamImposedLoad,
 		int beamSpacing,
 		@NotNull Set<Variant> variants
